refactor(core): extract DataFetcher implementations to controllers/data-fetchers
Move all inline DataFetcher implementations into the framework-agnostic
controllers layer for reusability across all framework bindings.

- Add mockDataFetcher (10k-line perf test generator)
- Add syntheticDataFetcher (preview app random data generator)
- Add baostockDataFetcher (real API adapter for stockbao backend)
- Remove inline mockDataFetcher from KLineChart.vue
- Remove inline synthetic dataFetcher from Vue/React preview apps
- Export all from @363045841yyt/klinechart-core/controllers
