+++ packages/core/src/controllers/data-fetchers/index.ts
@@ -1,0 +1,3 @@
+export { mockDataFetcher } from './mock'
+export { syntheticDataFetcher } from './synthetic'
+export { baostockDataFetcher } from './baostock'
