+++ packages/vue/preview/App.vue
@@ -116,7 +116,7 @@
 import { ref, computed, provide, inject, type Ref, type InjectionKey } from 'vue'
 import KLineChart from '../src/components/KLineChart.vue'
 import { VERSION, CORE_VERSION } from '../src/version'
-import type { DataFetcher } from '@363045841yyt/klinechart-core/semantic'
+import { syntheticDataFetcher } from '@363045841yyt/klinechart-core/controllers'
 import debugConfig from './debug-config.json'
 
 const FULLSCREEN_TARGET_KEY: InjectionKey<Ref<HTMLElement | null>> = Symbol(
@@ -137,42 +137,7 @@ function useFullscreenTeleportTarget() {
 const defaultConfig = debugConfig as Record<string, any>
 const currentConfig = computed(() => defaultConfig)
 
-const dataFetcher: DataFetcher = async (_source, config) => {
-  const start = new Date(config.startDate).getTime()
-  const end = new Date(config.endDate).getTime()
-  const dayMs = 86400000
-  const totalDays = Math.floor((end - start) / dayMs) + 1
-  const data: Array<{
-    timestamp: number
-    open: number
-    high: number
-    low: number
-    close: number
-    volume: number
-    turnover: number
-  }> = []
-  let price = 10 + Math.random() * 5
-  for (let i = 0; i < totalDays; i++) {
-    const ts = start + i * dayMs
-    const change = (Math.random() - 0.48) * price * 0.06
-    const open = price
-    const close = Math.round((open + change) * 100) / 100
-    const high = Math.round(Math.max(open, close) * (1 + Math.random() * 0.03) * 100) / 100
-    const low = Math.round(Math.min(open, close) * (1 - Math.random() * 0.03) * 100) / 100
-    const volume = Math.round(Math.random() * 10000000 + 1000000)
-    data.push({
-      timestamp: ts,
-      open,
-      high,
-      low,
-      close,
-      volume,
-      turnover: Math.round((volume * (open + close)) / 2),
-    })
-    price = close
-  }
-  return data
-}
+const dataFetcher = syntheticDataFetcher
 
 const displayVersion = `Vue@${VERSION}-Core@${CORE_VERSION}`
